Remove StompProtocol
