+++ stompman/__init__.py
@@ -1,4 +1,4 @@
-from stompman.client import Client
+from stompman.client import Client, ConnectionParameters, Heartbeat, Subscription, Transaction
 from stompman.connection import AbstractConnection, Connection
 from stompman.errors import (
     ConnectionConfirmationTimeoutError,
@@ -27,7 +27,6 @@ from stompman.frames import (
     SubscribeFrame,
     UnsubscribeFrame,
 )
-from stompman.protocol import ConnectionParameters, Heartbeat, StompProtocol, Subscription, Transaction
 
 __all__ = [
     "AbortFrame",
@@ -55,7 +54,6 @@ __all__ = [
     "NackFrame",
     "ReceiptFrame",
     "SendFrame",
-    "StompProtocol",
     "SubscribeFrame",
     "Subscription",
     "Transaction",
